datalib: templates cleaned to generate compilable code

+++ datalib/src/main/java/com/biorecorder/data/frame/impl/IntCount.java
@@ -11,6 +11,6 @@ class IntCount extends IntAggFunction {
 
     @Override
     protected int getValue1() {
-        return count;
+        return (int)count;
     }
 }
