+++ datalib/src/main/java/com/biorecorder/data/sequence/ShortSequence.java
@@ -5,6 +5,6 @@ package com.biorecorder.data.sequence;
  * that can be accessed but can not be modified
  */
 public interface ShortSequence {
-    public long size();
-    public short get(long index);
+    public int size();
+    public short get(int index);
 }
